+++ src/server/game/Server/Protocol/Handlers/BattleGroundHandler.cpp
@@ -338,7 +338,7 @@ void WorldSession::HandleBattlefieldListOpcode(WorldPacket &recv_data)
 {
     sLog->outDebug("WORLD: Recvd CMSG_BATTLEFIELD_LIST Message");
 
-    uint32 bgTypeId;
+    uint8 bgTypeId;
     recv_data >> bgTypeId;                                  // id from DBC
 
     uint8 fromWhere;
